requirements.txt
# NetPulse-Shield dependencies
# Install with:  pip install -r requirements.txt
ollama>=0.1.0
# Anomaly detection
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# RAG / LangChain (LCEL-based, compatible with LangChain 1.x)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# Embeddings (local, no API key needed)
sentence-transformers>=2.2.2

# Vector store
faiss-cpu>=1.7.4

# Email notifications
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# Dashboard
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0

# Optional: REST API
flask>=2.3.0
